Extract NoteContainer
